requirements.txt

python-telegram-bot==13.15
firebase-admin==6.4.0
